perf: defer gamer profile graph
